pyproject.toml
[project]
name = "helical-tech-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anndata>=0.12.10",
    "django>=6.0.3",
    "django-ninja>=1.6.2",
    "numpy>=2.2.6",
    "pandas>=2.2.2",
    "scipy>=1.13.1",
    "strawberry-graphql>=0.312.2",
    "strawberry-graphql-django>=0.82.0",
]

[dependency-groups]
dev = [
    "django-types>=0.23.0",
    "ipython>=9.11.0",
    "pandas-stubs>=3.0.0.260204",
    "scipy-stubs>=1.17.1.3",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["./src"]
include = ["logic", "server"]


[tool.basedpyright]
typeCheckingMode = "standard"
